requirements.txt
flask>=3.0.0
gTTS>=2.5.0
Pillow>=10.0.0
imageio>=2.30.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0
